Move Distribute to Structural Operations on Expressions
