+++ mathics/builtin/exp_structure/__init__.py
@@ -1,6 +1,8 @@
-"""
-Expression Structure
+r"""
+Structural Operations on Expressions
+
+Here we have functions which work purely on \Mathics3 Expressions transforming them in some way.
 """
 
 # This tells documentation how to sort this module
-sort_order = "mathics.builtin.expression-structure"
+sort_order = "mathics.builtin.structual-operations-on-expressions"
