+++ mathics/eval/exp_structure/__init__.py
@@ -1,0 +1,13 @@
+"Evaluation methods for builtin functions of mathics.builtin.exp_structure"
+
+from mathics.eval.exp_structure.distribute import eval_Distribute
+
+# from mathics.eval.exp_structure.outer import eval_Outer
+
+# TODO: add FlattenAt
+# from mathics.eval.exp_structure.flattenAt import eval_At
+__all__ = [
+    "eval_Distribute",
+    # "eval_FlattenAt",
+    # "eval_Outer",
+]
